@@ -1,10 +1,13 @@
 ﻿using System;
+using System.Threading.Tasks;
 
 namespace Shuttle.Recall.Tests
 {
     public class OrderHandler :
         IEventHandler<ItemAdded>,
-        IEventHandler<OrderSnapshot>
+        IEventHandler<OrderSnapshot>,
+        IAsyncEventHandler<ItemAdded>,
+        IAsyncEventHandler<OrderSnapshot>
     {
         private int _count;
         private DateTime _timeOutDate = DateTime.MaxValue;
@@ -25,6 +28,20 @@ namespace Shuttle.Recall.Tests
         public void Start(int handlerTimeoutSeconds)
         {
             _timeOutDate = DateTime.Now.AddSeconds(handlerTimeoutSeconds < 5 ? 5 : handlerTimeoutSeconds);
+        }
+
+        public async Task ProcessEventAsync(IEventHandlerContext<ItemAdded> context)
+        {
+            _count++;
+
+            await Task.CompletedTask;
+        }
+
+        public async Task ProcessEventAsync(IEventHandlerContext<OrderSnapshot> context)
+        {
+            _count++;
+
+            await Task.CompletedTask;
         }
     }
 }
